renamed MOTD file
