+++ packages/models/src/index.ts
@@ -1,1 +1,1 @@
-export * from "./MessageOfTheDay";
+export * from "./messageOfTheDay";
