+++ packages/models/src/messageOfTheDay.ts
